requirements/base.txt
-r openstack.txt
-r hermes.txt
Django==2.2.28
Faker==0.9.0
Markdown==3.0
Pillow==6.2.2
Unidecode==0.04.18
dj.choices==0.11.0
django-cryptography==0.3
django-extensions==2.2.9
django-filter==2.0.0
django-import-export==1.2.0
django-money==0.12.3
django-mptt==0.11
django-prometheus==2.2.0
django-reversion==3.0.5
django-rq==2.0
django-sitetree==1.13.0
django-taggit-serializer==0.1.7
django-taggit==0.22.2
django-threadlocals==0.8
djangorestframework==3.13.1
djangorestframework_xml==2.0.0
drf-nested-routers==0.92.5
factory-boy==2.11.1
mysqlclient==1.3.13
netaddr==0.7.18
openpyxl==2.4.0
py-moneyed==1.2
python-dateutil==2.4.2
pytz==2024.2
redis==3.2.1
requests-oauthlib==1.3.0
requests==2.20.0
rq==1.0
six>=1.9.0
sqlparse==0.2.3
tablib==0.11.5
typing==3.6.6
